+++ ST_engine/ST_audio_manager/src/test/audio_manager_tests.cpp
@@ -16,19 +16,19 @@ class audio_manager_test : public ::testing::Test {
 
 protected:
 
-    void set_chunks(ska::bytell_hash_map<size_t, Mix_Chunk*>* assets){
+    void set_chunks(ska::bytell_hash_map<uint16_t , Mix_Chunk*>* assets){
         test_mngr->chunks_ptr = assets;
     }
 
-    void set_music(ska::bytell_hash_map<size_t, Mix_Music*>* assets){
+    void set_music(ska::bytell_hash_map<uint16_t, Mix_Music*>* assets){
         test_mngr->music_ptr = assets;
     }
 
-    void play_sound(size_t arg, uint8_t volume, uint8_t loops){
+    void play_sound(uint16_t arg, uint8_t volume, uint8_t loops){
         test_mngr->play_sound(arg, volume, loops);
     }
 
-    void play_music(size_t arg, uint8_t volume, uint8_t loops){
+    void play_music(uint16_t arg, uint8_t volume, uint8_t loops){
         test_mngr->play_music(arg, volume, loops);
     }
 
@@ -58,7 +58,7 @@ protected:
 TEST_F(audio_manager_test, test_play_wav_full_volume) {
     auto test_wav = Mix_LoadWAV("test_sound.wav");
     ASSERT_TRUE(test_wav);
-    ska::bytell_hash_map<size_t, Mix_Chunk*> assets;
+    ska::bytell_hash_map<uint16_t, Mix_Chunk*> assets;
     assets[1] = test_wav;
     set_chunks(&assets);
     play_sound(1, MIX_MAX_VOLUME, 0);
@@ -69,7 +69,7 @@ TEST_F(audio_manager_test, test_play_wav_full_volume) {
 TEST_F(audio_manager_test, test_play_wav_half_volume) {
     auto test_wav = Mix_LoadWAV("test_sound.wav");
     ASSERT_TRUE(test_wav);
-    ska::bytell_hash_map<size_t, Mix_Chunk*> assets;
+    ska::bytell_hash_map<uint16_t, Mix_Chunk*> assets;
     assets[1] = test_wav;
     set_chunks(&assets);
     play_sound(1, MIX_MAX_VOLUME/2, 0);
@@ -79,7 +79,7 @@ TEST_F(audio_manager_test, test_play_wav_half_volume) {
 TEST_F(audio_manager_test, test_play_wav_looping) {
     auto test_wav = Mix_LoadWAV("test_sound.wav");
     ASSERT_TRUE(test_wav);
-    ska::bytell_hash_map<size_t, Mix_Chunk*> assets;
+    ska::bytell_hash_map<uint16_t, Mix_Chunk*> assets;
     assets[1] = test_wav;
     set_chunks(&assets);
     play_sound(1, MIX_MAX_VOLUME, 3);
@@ -89,7 +89,7 @@ TEST_F(audio_manager_test, test_play_wav_looping) {
 TEST_F(audio_manager_test, test_play_ogg_full_volume) {
     auto test_ogg = Mix_LoadMUS("test_music.ogg");
     ASSERT_TRUE(test_ogg);
-    ska::bytell_hash_map<size_t, Mix_Music*> assets;
+    ska::bytell_hash_map<uint16_t, Mix_Music*> assets;
     assets[1] = test_ogg;
     set_music(&assets);
     play_music(1, MIX_MAX_VOLUME, 3);
@@ -99,7 +99,7 @@ TEST_F(audio_manager_test, test_play_ogg_full_volume) {
 TEST_F(audio_manager_test, test_play_ogg_half_volume) {
     auto test_ogg = Mix_LoadMUS("test_music.ogg");
     ASSERT_TRUE(test_ogg);
-    ska::bytell_hash_map<size_t, Mix_Music*> assets;
+    ska::bytell_hash_map<uint16_t, Mix_Music*> assets;
     assets[1] = test_ogg;
     set_music(&assets);
     play_music(1, MIX_MAX_VOLUME/2, 3);
